+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidMemoryCollector.java
@@ -1,17 +1,26 @@
 package io.sentry.android.core;
 
 import android.os.Debug;
-import io.sentry.IMemoryCollector;
+import io.sentry.ICollector;
 import io.sentry.MemoryCollectionData;
+import io.sentry.PerformanceCollectionData;
 import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
 
 @ApiStatus.Internal
-public class AndroidMemoryCollector implements IMemoryCollector {
+public class AndroidMemoryCollector implements ICollector {
+
   @Override
-  public MemoryCollectionData collect() {
+  public void setup() {}
+
+  @Override
+  public void collect(@NotNull Iterable<PerformanceCollectionData> performanceCollectionData) {
     long now = System.currentTimeMillis();
     long usedMemory = Runtime.getRuntime().totalMemory() - Runtime.getRuntime().freeMemory();
     long usedNativeMemory = Debug.getNativeHeapSize() - Debug.getNativeHeapFreeSize();
-    return new MemoryCollectionData(now, usedMemory, usedNativeMemory);
+    MemoryCollectionData memoryData = new MemoryCollectionData(now, usedMemory, usedNativeMemory);
+    for (PerformanceCollectionData data : performanceCollectionData) {
+      data.addMemoryData(memoryData);
+    }
   }
 }