+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add cpu usage collection ([#2462](https://github.com/getsentry/sentry-java/pull/2462))
 - Improve ANR implementation: ([#2475](https://github.com/getsentry/sentry-java/pull/2475))
   - Add `abnormal_mechanism` to sessions for ANR rate calculation
   - Always attach thread dump to ANR events